Text To Speech Module
This PR adds a Text-To-Speech Module to Expo.

It enables using built-in TTS functionalities - the user can choose a language, voice rate and pitch that will be used to speak out loud the passed text.

Every invocation of  will add an utterance to queue - anyway, the user has also an ability to stop the speaking. The module also exposes a method to check if speaking is now in progress.

What'\''s more every speech has progress listeners attached : when the utterance is , ,  by the user or an  occurred (Android only). Listeners'\'' callbacks can be passed via the  parameter.

As long as iOS AVSpeechUtterance has no identificator I had implement a subclass that has one to have an ability to identify which utterance has started/finished/etc.

Closes expo/expo#311

fbshipit-source-id: 651853234c