Use os_log in CocoaLumberjack on iOS 10+
os_log is the new Apple Unified Logging API introduced in iOS 10. I have a follow-up PR that uses it for logging test results. For os_log calls to be enabled, we need to remove the OS_DISABLE_ACTIVITY env var, which was set previously because Xcode 7 (?) had a ton of log spew without it back in the day, and that's since been fixed.

Enabling os_log messages results in double-printing of JS `console.X` messages, though, since we configured RN to use CocoaLumberjack and we configured CocoaLumberjack to write to ASL (main logs) and TTY (Xcode) loggers. This is no longer necessary as of iOS 10 -- CocoaLumberjack just needs to write to os_log.

Tested by making sure os_log calls were visible in both Xcode and when streamed with `xcrun simctl spawn booted log stream`.

fbshipit-source-id: c884aa7