Allow expo to run on Android TV
Hi,

As I am currently developing an Android TV app, I forked your project to build a Android-TV compatible version of your Android client.

It was a few lines to add to the AndroidManifest and a new activity.

I have tested it on my Android TV and it worked well !

Have a nice day,

Maxence

Closes expo/expo#1957

fbshipit-source-id: dfca5829