[sdk31][tests] Use babel.config.js instead of .babelrc in the root to get Home and Expo tests passing
This replaces .babelrc in the workspace root, which is good because we have different types of projects in this repo and they shouldn't share the same Babel configuration necessarily. Instead, we now have a babel.config.js file that belongs in each individual workspace that has tests.

I also had to add a postinstall script to the workspace root to delete `node_modules/babel-jest/node_modules/babel-core`, which was inexplicably installed even though babel-jest does not depend on babel-core, even transitively (Yarn must be buggy). I also added `babel-core` v7 to the workspace root to force to get hoisted so that babel-jest would load it. This reaffirms the fragility of Babel and Jest and why we need to minimize the projects in this repo and be explicit and deliberate when we update these dependencies.

Test plan: Ran `yarn test` with `--no-cache` (to ensure Babel runs) in home and packages/expo; all tests pass