[vscode] Add shared VS Code settings (to use TypeScript from npm)
This adds the VS Code settings file, which means that everyone developing in this repository with VS Code will use the same settings, overlaid on top of their personal global settings. This ensures we can set things like line widths, spaces, and what version of TS to use in a consistent way.

This commit tells VS Code to use TypeScript from the workspace root instead of whatever the TypeScript plugin has installed. This is so that TS messages in the editor match `tsc`'s messages from the CLI.