[yarn] Add a script to check the current version of Yarn (#2314)
Since we aren't using Nix to control the version of Yarn, in practice we've seen several people have different versions of Yarn installed -- I've seen this twice already. This commit adds a script called `check-yarn` that looks at your version of Yarn and compares (using semver) its version against a minimum desired version that is hard-coded.

The script runs as a "preinstall" script and not in `.envrc` because `yarn --version` is a bit slow and takes about 250ms on my machine and making `cd` feel noticeable is annoying.

Also actually enabled `source_local` in `.envrc` for good measure.

Test plan: Uninstalled the latest version of Yarn and ran `yarn` in the repository root and got the warning message about my Yarn version. Upgraded Yarn and saw the message go away.