[test] Upgrade ts-jest in the expo package
This upgrade ts-jest to the latest version, which has a slightly different config format. I think we want to keep it separate from the jest-expo preset so that if ts-jest becomes unsupported or if people rely on ts-jest as a public feature we're not stuck with it, but it would be nice to add this to our common scripts eventually so that all unimodules in this repo can use it.

Test plan: run tests with `yarn jest --no-cache` to force rebuilds.