[video] Move references to static properties outside of the Video component to fix HMR transform bug
The HMR transform is now applied always in dev and acts as a decorator, which breaks static properties on components. To get around this, I moved the static properties out of the class.

Test plan: opened the Video demo in NCL and verified the video now covers the entire visible area.