[expotools] Add `expotools`/`et` to bin
Added convenience commands based off of the setup in Universe. There are two scripts named `expotools` and `et` that do the same thing. They use Nix to build and track a copy of Expotools so that if there are any changes to Expotools, we display a prompt to the developer on the next run so that everyone is running the correct copy of Expotools for their checkout of the repo.

The Nix package for expotools ignores the `build` and `node_modules` directories because Nix creates those build artifacts itself when creating the package. To ignore those directories, we use a negated version of `lib.sourceByRegex`.

Test plan: run `expotools` in the Expo repo and get a prompt to build Expo tools. Then run `yarn` in expotools to build it normally, without Nix, which generates the build and node_modules directories. Run `expotools` again to verify Nix reuses the old expotools without letting those directories bust the Nix cache.