[nix] Install `expo-cli` without optional dependencies
To ensure determinacy, node2nix treats these the same as dependencies,
meaning that unlike when using npm or yarn, the build of a package
will fail if any of its optional dependencies fail to build.

`expo-cli` depends on traveling-fastlane-linux and -darwin, at least
one of which will always fail to install.

Within this repo we can ensure fastlane itself will also be installed
whenever the expo cli is.