[ts] Add "dom" to the set of TS libraries
This makes it so that we don't get type errors -- and actually get autocompletion and checking -- fwhen accessing browser APIs for Expo Web.

This has the downside of making it seem like browser APIs are available outside of the browser like on native so we still need other tests to cover that.

Test plan: `yarn build` in the expo package