[ios] add super calls to template AppDelegate.m (#3275)
# Why

Follow-up to #3169 

After discussion on slack with @tsapeta , we decided the current template AppDelegate.m file makes it difficult and confusing to install third-party RN libraries that also require editing AppDelegate. (Developers would need to remember to add a call to `super` into specific methods, and the failure case isn't great -- something small would just break.) To remedy this, we decided to explicitly add in the methods that need to call `super`.

# How

Made stubs that call `super` for all of the methods that are defined in `EXStandaloneAppDelegate.m`. Also edited docs a bit to clarify instructions in light of this change.

# Test Plan

- [x] Make an iOS ExpoKit app with this AppDelegate and make sure it builds and runs properly
- [x] Test opening the app from a URL, make sure super is being called