[docs] Add redirect for installation page + update link in Home
The link https://docs.expo.io/versions/latest/introduction/installation.html no longer exists but is linked to from the Expo client. This commit changes that link in Home to be https://docs.expo.io/versions/latest/introduction/installation/ and adds a redirect on the docs site so that this old link still works for now.