Add more issue templates (#3761)
# Why

It becomes annoying when we get new issues that contain fragments not removed from the template and have `READ THIS FIRST, PLEASE!` header 😅 
A few recent examples: #3760, #3728, #3714

# How

I utilized GitHub issue types and the ability to have multiple issue templates. When creating a new issue, users will be asked to choose a type of the issue and depending on that, different issue template will be used. Some of them will be redirecting users to other repositories like `expo-cli` or pages like canny or forums.

It'll be great if we could also add a bot that will be automatically closing issues for cli, feature requests and discussions 🤔 

# Test Plan

We will see once merged 😁