[ios] Add some unprefixed macros to the versioning regexes
The macros ENABLE_PACKAGER_CONNECTION (in react-native) and HAVE_GOOGLE_MAPS (new in react-native-maps) were not versioned. This commits makes the versioning script add the ABI prefix to these macros.

Also made the versioning script delete JS files, since we don't need those in the versioned code. This means fewer files added to Git and we won't need to blacklist the iOS versioned directories for Metro.

Tested with `gulp ios-add-version --abi 33.0.0 --root ..` from the tools directory. Grepped for the macros in the generated files and saw that they were properly namespaced.