Enable User Defined Build Settings in Release Channel (#2093)
Xcode's User Defined Build Settings can (unfortunately) only be used in Info.plist so this allows you to set Expo's Release Channel there. When you're detached, this is helpful in order to automate different Release Channels for dev, staging, prod. Additionally, when you're detached, the Version and Build of the binary should generally be part of release channel as well. Version and Build can be stored in User Defined Build Settings and used to set both CFBundleShortVersionString, CFBundleVersion and as part of ExpoReleaseChannel to avoid duplication.