[home] Revert app.json changes from 4b47e08ac0 so we can publish prod home
Home's app.json needs special fields in it so that Expo CLI knows where to save the published bundle that gets embedded. This commit reverts the changes to app.json made in 4b47e08ac0. I also formatted the file so that it's valid JSON + comments, which VS Code is better at handling.