[home][important] Clear the local session secret if it is no longer valid
If the local session secret is not valid and the server returns successfully but gives `null` for the viewer, this means the viewer is not logged in. In `MyProfileContainer` we expect the user to be logged in (we render this component only when the local session secret is set) so if we get back `null` this means the session secret is no longer valid (ex: was revoked).

Test plan: Log in, then go to expo.io and log into the same account and log out of all other sessions, then refresh the profile screen on Home and ensure we are taken back to the login screen.

Also ensured that we do only one logout operatio since Apollo will re-render the container twice -- once when making the request and once again when the response comes back -- we want to do the check only when the response is complete and there was no error.