[expo] Remove most expo-* packages from expo package (#4815)
# Why

In SDK33 we deprecated imports from the Expo package, instead users install them separately, eg: `expo install expo-sensors`. This has many benefits that we're already aware of that I'm not going to repeat here for brevity. 

# How

I removed the dependencies and tested in a local bare project to ensure that the project would build and run in Xcode and also that it would work in the Expo client.

I had to keep some of the unimodule dependencies around because they are used by parts of the expo package in a way we can't remove (globals, for example).

# Test Plan

You can install `expo@34.0.0-experiment.6` to a bare project and give it a try!

# Checklist

- [x] Verify it works as expected in bare on iOS and Android
- [x] Verify it works as expected in managed on iOS and Android
- [x] Verify it works for web
- [x] Add docs for community native modules we include now that they can't be imported from the expo package
- [x] Create a package for the `Google.ts` file so it doesn't need to live in the `expo` package and force a dependency on `expo-app-auth`
- [x] Update new project template to include RNGH

# Followup

- @EvanBacon is planning on removing maps related code from the package and into something like `expo-maps`