[calendar] fix Alarm's relativeOffset's type (#4905)
According to the docs the correct type is number though here it was
string. This commit fixes that.

Docs: https://docs.expo.io/versions/latest/sdk/calendar/#alarm

# Why

Typing is wrong.

# How

Changed the type.

# Test Plan

Try using it, it doesn't work. It expects a number but the type in typescript is a string.