[ios13] Add support for iOS 13's unused attribute marker
iOS 13 changed the unused parameter attribute to `__attribute__((__unused__))`. We need to support this in `RCTParseUnused` to prevent a crash when the app is built with Xcode 11 and run on an iOS 13 simulator.

We'll backport this to the SDK 34 branch so that ExpoKit (but not the iOS client, which is already frozen) will get this fix -- no more errors about "Unknonw argument type `__attribute__` in method".

Test plan: compile the Expo client with Xcode 11 and run it in an iOS 13 simulator and run the client successfully.