[lint] Lint the `expo` package with `expo-module lint`
Fixes lint errors and warnings in the `expo` package. We use `expo-module lint`, which is designed for this repo. It assumes that TS/JS source is under `src` and will lint that directory. The base ESLint config uses `universe/native` but adds in the Node environment for tests, since they run in Jest.