expo-contacts: fix getting containers without a predicate (and crash) (#6016)
The predicate was not initialised in all cases, which meant that an invalid
predicate passed from javascript would make it crash.

Additionally, according to the Apple's docs[1], passing nil as the predicate
is fine and just means no filter. At the moment there's no way to get the
whole list of containers without filtering. This patch fixes that.

1: https://developer.apple.com/documentation/contacts/cncontactstore/1403238-containersmatchingpredicate?language=objc