[gl-cpp] Fix type problem with EXiOSOperatingSystemVersion (#6063)
# Why

This line prevents build on XCode in a bare react project (at least). Because line 12 of the same file uses `EXiOSOperatingSystemVersion` instead of `struct EXiOSOperatingSystemVersion`. 

# How

Since other files of the same package use `EXiOSOperatingSystemVersion` and not `struct EXiOSOperatingSystemVersion`, I created a typedef, which solved the problem

# Test Plan

It builds on XCode (at least)