[sdk36][gl] Fix expo-gl crashing instead of warning when initialized in remote JS session (#6381)
# Why

Issue found when testing #6372 - initializing a GL context is crashing Expo client (`EXC_BAD_ACCESS`) when remote JS debugger is on.
Also, it may change the behavior that some people were reporting here: #5493, but it doesn't fix it completely, just prevents crashes.

# How

`*(((JSGlobalContextRef *)(_bridge.runtime)) + 1)` - this hack caused `EXC_BAD_ACCESS` crash which cannot be try-catched but we can check whether `_bridge.runtime` exists and is truthy before running such code. If not, we can just return `nil` which means that JS remote debugging is enabled. (In that case expo-gl warns that it doesn't work in such session)

# Test Plan

Tested in NCL with remote JS debugger enabled.