[expo-notifications] Implement a simple ExpoNotificationsService (#7068)
# Why

We want to be able to handle notifications even when the app is running — this we will achieve using a `Service`.

# How

To be able to present notifications when the app is not running we need access to a builder. In order to implement this I moved the builder out from the unimodules structure and added a direct dependency on it to `ExpoNotificationsService`.

I imagine `ExpoNotificationsService` to become the entrypoint for all the notifications. If a developer needs to extend the building process anyhow, they will be able to subclass `ExpoNotificationsService` and override `getNotificationBuilder` function, or even `onNotificationPresent`. `android:order="-1"` attribute will ensure their service will always be resolved with higher priority.

The overall architecture will look like this (semi-transparent elements are not implement as of yet):

![excalidraw-202012311929-20](https://user-images.githubusercontent.com/1151041/74679537-6c402280-51be-11ea-8e97-d2df9bea181e.png)

Before the rearchitecture it looked more like this:

![excalidraw-202012311929-11](https://user-images.githubusercontent.com/1151041/73084552-7461b680-3ecd-11ea-856c-89fb3d59b95c.png)


# Test Plan

`test-suite` passes, notifications are shown.