[expo-notifications] Handle notifications when app is in background (#7213)
# Why

The fact that the app is in background or closed shouldn't prevent us from presenting the notification.

# How

`ExpoNotificationsService`, the central dispatcher of the notifications, listens for activity state changes and when the app is not in foreground and a notification received it is presented with `null` behavior. This makes the builder fall back to using notification-payload values (as iOS does).

# Test Plan

Added an interesting test case which requires user interaction. It is automatically skipped after 5 seconds, so it won't mess with the test result. The test passes on both Android and iOS.

Moreover, both a scheduled repeating notification and a remote notification are shown when the app is closed.