[gl-cpp] Fix depth/stencil buffers not working correctly with three.js (#7543)
# Why

Fixes #7502 

# How

It turned out that Three.js uses `gl.DEPTH_STENCIL` constant when calling `renderbufferStorage` method used as part of lights rendering. This constant is valid in WebGL, but is invalid in OpenGL ES 3.0, so we need to fall back it to `GL_DEPTH24_STENCIL8` which is valid in OpenGL and has the same effect.

# Test Plan

Added new example to NCL which is almost the same as the one provided in the issue. Also received a feedback that JS workaround doing the same thing works as expected.