[expo] Move font-family preprocessor to top-level scope (#7805)
# Why

Followup #7800 - after applying changes in that PR, there is still an issue with vector icons not being rendered correctly. That happens only when the experience is launched through the deep link and client app was killed. It this scenario, the `main` root component of home app wasn't even created, so `StyleSheet.setStyleAttributePreprocessor('fontFamily', Font.processFontFamily);` was never called, resulting in vector icons being rendered in the dev menu as non-existing characters.

# How

In my opinion, this preprocessor should be set once and not only in the root Expo component, so top-level scope fits better.

# Test Plan

I've tested whether it solves described issue in the dev menu - it does. I've run home and ncl locally and I didn't notice any new issues.