[tools-public] Remove check-dynamic-macros-android.sh (#7919)
# Why

Follow-up to https://github.com/expo/expo-cli/pull/1933.

# How

Since we won't use `check-dynamic-macros-android.sh` when building standalone apps and there is no other place we use it we can remove it. Let's believe in `expotools`!

# Test Plan

I will cherry-pick that commit to `sdk-37` and push a new image to Android staging.