[expo-notifications] Remove need for dependency on newer @unimodules/core (#8162)
# Why

Adding a dependency on newer  (`>= 5.1.1`) `@unimodules/core` breaks installations with `react-native-unimodules@0.9.0` which depends on `unimodules-app-loader@1.0.1` which is incompatible with newer versions of `@unimodules/core`. Making `expo-notifications` backwards compatible is the easiest way to fix this complex tie of dependencies.

# How

Since the `@unimodules/core@5.1.0` incompatibility comes from inability to handle `null` values in `Map`s passed to `MapArguments`, we can guard ourselves against that scenario by catching the `NullPointerException` and stripping out the invalid values when in need.

# Test Plan

I have confirmed that if `ReadableArguments` do not handle `null` values this code catches the exception properly and succeeds at serializing the notification.