[gl] Add WebP format as an option when taking GL snapshots (Android only) (#7490)
# Why

At some point (API 14), Android added native WebP support ([Bitmap.CompressFormat.WEBP](https://developer.android.com/reference/android/graphics/Bitmap.CompressFormat#WEBP)), so adding the option to export the render to WebP is trivial.
Creating a new module just for converting from PNG to WebP is silly since you can get the GL render as WebP directly. Also converting from PNG to WebP could create quality loss.

# How

In GLContext.java, I extended the block that specifies using png (instead of jpeg, the default) to recognize webp as another posible format.

# Test Plan

Sadly, a combination of living in Windows OS and lack of experience in some aspects of software development, I was not able to test this change using the proper methods (unit test, and that). But I did tried the change in the App I'm building and it worked fine

Thanks for your attention, sorry for any mistake in the PR, this is the second time I do a PR