[expo-notification] Adds DailyTrigger (#8199)
# Why

In order to trigger a local notification once per day, the `DailyIntervalTriggerInput` is being added with this PR.
Previously there was no solution that would allow this with the expo-notification library. It was possible on iOS but not Android. 

# How

The daily interval uses the `TimeIntervalTrigger` on Android with a modified subsequent day trigger interval and on iOS it uses UNCalendarNotificationTrigger.

# Test Plan

Added two integration tests but those tests only verify first day event fire. We may want to remove these are they both take over 1 minute to run. If there are better solutions to verify the notification was scheduled properly, please let me know in the comments.