[expo-image] Clipping, shadows, borders & background-color (#7457)
# Why

This PR adds support for Clipping, shadows, borders & background-color to `expo-image`.

# How

## iOS

In order to support proper clipping & shadows, the component has been changes to a regular `UIView` with a `UIImageView` as a child. This makes it possible to support shadows correctly by clipping the contents of the child `UIImageView`.

Borders are implemented using 1 or more layers. Layers are a very efficient and highly optimised rendering technique and are animatable. When all border-edges are visible (and have the same width/color/style), a single layer is used as an optimal rendering path. When the borders are different, each edge gets its own layer, clipped to that edge.

## Android

On Android, the clipping and shadow drawing are performed by the `OutlineProvider`. This will use hardware acceleration for rendering and clipping where possible. In some cases (when the border-radii are different), the code falls back to Canvas clipping, which is also pretty fast but doesn't support Anti-aliasing. Visually it very hard to spot the difference.

Border-drawing is performed by a "patched" version of the `ReactViewBackgroundDrawable` class. The applied patches only fix certain drawing issues but don't modify the API of `ReactViewBackgroundDrawable`. This makes it possible to switch back to using the original `ReactViewBackgroundDrawable` class, once all the fixes have landed upstream in react-native.

# Test Plan

Additional tests have been added to NCL to verify the output.

![ios-shadows](https://user-images.githubusercontent.com/6184593/77431379-e88cdd80-6ddc-11ea-8e48-00ed02b2b82e.png)
![android-shadows](https://user-images.githubusercontent.com/6184593/77431401-f0e51880-6ddc-11ea-8cf0-c2138d8ee338.png)
![ios-borders](https://user-images.githubusercontent.com/6184593/77431453-02c6bb80-6ddd-11ea-92c6-b57015fef253.png)
![android-borders](https://user-images.githubusercontent.com/6184593/77431464-065a4280-6ddd-11ea-91b9-213e43ae8ef6.png)


# Open issues

- on iOS, borders with different widths are currently not drawn correctly.

![image](https://user-images.githubusercontent.com/6184593/77430985-54bb1180-6ddc-11ea-87e8-61091c5ae48a.png)


- on Android, borders with different colors and different border-radii may result in drawing artefacts. This is an issue of `ReactViewBackgroundDrawable` class.

![image](https://user-images.githubusercontent.com/6184593/77430780-fee66980-6ddb-11ea-8c50-479ed49926e7.png)