[tests][detox] Make Relapse serialization more robust (#5765)
Sometimes unserializable values are passed into proxied methods. When this happens, catch the error and send an explicit message to the server where we can print out more information about what failed.

This commit adds an optional "code" field to the JSON payload sent across the web socket. There is one value -- RelapseCode.SerializationError -- that is supported. When it is sent, Jest prints out:
```
  console.error relapse/server.js:29
    Client failed to serialize the arguments for a call to console.assert
```