[ios14] Get the app store client compiling & running on iOS 14 (#8938)
The main compiler-breaking change was the addition of a new notification option ("ephemeral notifications" for App Clips) that we needed to declare in a switch-case statement. Also left some comments for where we want to use some new enum values that replace alerts.

# Test Plan
The app compiles for iOS 14 with Xcode 12 beta. I needed to run it twice -- the first time it got stuck on the launch screen with the spinner -- and was able to log into my account and load published NCL. Images don't show up.