[expo] Fix logs showing up in both consoles when debugging remotely (#8807)
# Why

Finally fixes the long-standing issue https://github.com/expo/expo/issues/2623.

# How

Fixed the `if` that decided whether to enable remote logging or not.

# Test Plan

It turns out JS, when debugged remotely, is run in a Worker, which exposes `WorkerNavigator` under the `navigator` property and it doesn't have the `userAgent` property as its own (the only two are `product` and `geolocation`, at least in my Chrome). We can't rely on `product`, as `navigator.product` is the same in debugged and not-debugged RN (`"ReactNative"`), so I fell back to an old `atob` check, suggested in https://stackoverflow.com/a/42839384/1123156 and confirmed to be working https://github.com/expo/expo/issues/2623#issuecomment-441364587.