[lint] Turn off react-hooks/exhaustive-deps by default (#9130)
This lint rule has been relatively noisy since it errs on the side of reporting false positives. Our linter philosophy is that needing to write `// eslint-ignore` indicates an issue with the linter choices and we err on the side of trusting the programmer.

It's easy to turn this back on in an app by adding `'react-hooks/exhaustive-deps': 'warn'` to the app's own config.

Removed the `// eslint-ignore` comments for this rule from all the places we had them. Updated the test snapshots and verified the changes looked right.

Ran the linter config unit tests and made sure they passed.