[expo-updates] run reaper inside of LoaderTask (#9767)
# Why

When I created the LoaderTask class in https://github.com/expo/expo/pull/9144, I left the reaper as something that the calling class should be in charge of. However, I've since realized that it's something the LoaderTask must do because only the calling class does not always have enough information to determine when it's safe to run the reaper. The reaper is also now smart enough to use the scopeKey from the passed configuration object and only reap updates for the given scopeKey, so this makes it safe for LoaderTask to run.

# How

Add calls to `runReaper()` in the same places they were before https://github.com/expo/expo/pull/9144, and remove them from the LoaderTask callbacks in UpdatesController. No changes necessary in the development clients.

# Test Plan

Load 2 OTA updates (for bonus points, use the development clients!) and look in the SQLite database to ensure that the first is reaped after the second one successfully launches. Also ensure that updates with a different scopeKey are NOT reaped.