[expo-document-picker] Handle dismissal by gesture well (#10325)
# Why

Fixes second point of https://github.com/expo/expo/issues/10303.

# How

Figured out neither of the lifecycle callback methods is called if the VC is dismissed by gesture. Tried to disable the gesture with `isModalInPresentation = true`, but that didn't work. Found a way to get a callback called when the VC is dismissed by gesture. Implemented it, figured it's working, backported to SDK39.

# Test Plan

I have tested manually that opening a document picker and dismissing it by dragging it down is caught and handled by the module.