[expo-notifications] Move handling implementation to NotificationsService delegate (#10666)
# Why

A prerequisite to single-threaded notifications handling.

# How

More refactoring, moving staff from `NotificationsHelper` to appropriate `NotificationsService`' delegates. This time to `ExpoHandlingDelegate`.

In the meantime I have also removed a memory leak introduced in https://github.com/expo/expo/pull/10640 — we don't need to _observe_ the state, we can check for the state every time we need to access it.

# Test Plan

Push and scheduled notifications are presented.