[ci] Fix NCL website deploy workflow
The secrets context is not available in if fields,
so before this commit the workflow would not run for
anyone.

After this commit, the workflow runs for all PRs that effect the related
files, but does not run the deploy steps if the PR author does not have
access to repository secrets.