[expo-updates][database] add json_data table (#11049)
# Why

In order to support the new EAS update manifest format and particularly the `serverDefinedHeaders` and `manifestFilters` fields, we need to be able to store some general JSON data from the server on disk (and then use it when requesting/launching future updates). It makes the most sense to use the SQLite database for this since we already use this to persist all other updates-related metadata.

# How

This PR adds a `json_data` table for storing arbitrary JSON blobs keyed by a string (like `manifestFilters`) and a scopeKey. Critically, this PR also bumps the updates SQLite database version number; this currently causes a destructive migration and requires that it be released in a major SDK release.

A future PR will use this table in the implementation of the `serverDefinedHeaders` and `manifestFilters` from the RFC.

# Test Plan

Built Expo client on both platforms and viewed the updates SQLite database in a desktop SQLite browser to ensure the table was created properly.