[expo-dev-launcher][Android] Apply user interface style (#11322)
# Why

Applies the user interface style from the app.json when the app is loaded on iOS.
It's similar to https://github.com/expo/expo/pull/11277.

# How

We need to change the local variable in the `AppearanceModule`. Unfortunately, it's the only solution, because the `AppearanceModule` normally reads the system settings which we can't change. 
It's worth mentioning that we can't change the theme of the Android UI - the user should force it by his `AndroidManifest.xml` (by Android UI, I meant, for example, data picker or image picker).  

Moreover, this won't work with the `expo/react-native-appearance` module - https://github.com/expo/react-native-appearance.

# Test Plan

- bare-expo ✅