[av][ios] Reduce time to record for Audio recordings (#12294)
Recording Audio on iOS is pretty slow to start on real devices and may take up to 1..1.5 seconds before it starts. This is caused in large part due to the AVSession which is deactivated and reactivated in between the prepare and start calls. This PR introduces an optimised API (similar to Sound.createAsync) to streamline audio recording. On iOS it sets a new option called keepAudioActiveHint which hints the native layer that start will be called immediately after and thus the session does not need to be deactivated.