[docs] convert next batch of modules to use auto generated API section (#12408)
# Why

This PR is the continuation of #12239, and includes next batch of modules converted to leverage the newly added auto generated API section feature.

# How

The new modules which will use this feature are:
* `expo-application`
* `expo-keep-awake`
* `expo-splash-screen`

I had to add the missing documentation for two methods in `Application` module, so this part of changeset should be verified by someone more knowledgable. Those methods were missing in the current docs.

The changes also have been backported to the SDK v41 docs (form `unversioned`, as in the origin PR), so the changes can be seen after the deploy.

# Test Plan

I have been testing and improving the autogen feature using those modules too, but they did not land in the initial PR to reduce the amount of files and changes in there.

All the pages have been tested on `localhost` and should include all the information which were previously present in thew docs with some additional tweak, fixes and small additions.